requirements.txt
asgiref==3.7.2
awsebcli==3.20.10
blessed==1.20.0
boto3==1.20.50
botocore==1.23.54
cement==2.8.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.3
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.58.1
django-braces==1.14.0
django-cors-headers==4.3.0
django-filter==23.5
django-rest-auth==0.9.5
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docutils==0.16
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
jmespath==0.10.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.10.1
Pillow==10.1.0
pyasn1==0.5.0
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.5.2
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.1.9
